Trello 2170 - split Reports.jsx into multiple subcomponents
* refactoring

* Code optimization

* Code optimization_3

* Code optimization

* Code optimization

* Code optimization

---------
